+++ library/LibUtilities/Memory/NekMemoryManager.hpp
@@ -187,7 +187,7 @@ public:
     /// instead.  Any memory allocated from this method must be returned to the
     /// memory pool via RawDeallocate.  Failure to do so will result in memory
     /// leaks and undefined behavior.
-    static DataType* RawAllocate(unsigned int NumberOfElements)
+    static DataType* RawAllocate(size_t NumberOfElements)
     {
 #ifdef NEKTAR_MEMORY_POOL_ENABLED
         return static_cast<DataType*>(GetMemoryPool().Allocate(sizeof(DataType)*NumberOfElements));
@@ -203,7 +203,7 @@ public:
     ///
     /// This method is not meant to be called by client code.  Use Array instead.
     /// Only memory allocated via RawAllocate should be returned to the pool here.
-    static void RawDeallocate(DataType* array, unsigned int NumberOfElements)
+    static void RawDeallocate(DataType* array, size_t NumberOfElements)
     {
 #ifdef NEKTAR_MEMORY_POOL_ENABLED
         GetMemoryPool().Deallocate(array, sizeof(DataType)*NumberOfElements);
